+++ frontend/e-commerce-service/src/app/components/error-pages/page-not-found/page-not-found.component.ts
@@ -12,7 +12,7 @@ export class PageNotFoundComponent {
 
   constructor(private router: Router) {}
 
-  handleLogoClick(): void {
+  handleLogo(): void {
     this.router.navigate(['/']);
   }
 }