+++ frontend/e-commerce-service/src/app/components/header/header.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef, EventEmitter, Input, OnInit, Output, Renderer2, ViewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
 import { AuthService } from '@auth0/auth0-angular';
@@ -7,6 +7,7 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import { faList, faSearch, faShoppingCart, faSignIn } from '@fortawesome/free-solid-svg-icons';
 import { CartResponseDTO, CartService } from '../../generated';
+import { ProductListFilter } from '../../models/product-list-filter';
 
 @Component({
   selector: 'app-header',
@@ -21,11 +22,23 @@ export class HeaderComponent implements OnInit {
   public faSignIn: IconDefinition = faSignIn;
   public faList: IconDefinition = faList;
 
-  public searchText: string = ""; 
   public isLoggedIn: boolean = false;
   public accessToken: string | null = ''; 
 
-  constructor(private router: Router, public auth: AuthService, private cartService: CartService) {}
+  // filter and pagination properties
+  @Input() enableSearch: boolean = true;
+  @Input() availableCategories: string[] = [];
+  @Output() searchChanged = new EventEmitter<ProductListFilter>();
+  @ViewChild('collapseElement') collapseElement!: ElementRef;
+
+  public productListFilter: ProductListFilter = {
+    category: null,
+    minPrice: null,
+    maxPrice: null,
+    searchText: null
+  };
+
+  constructor(private router: Router, private renderer: Renderer2, public auth: AuthService, private cartService: CartService) {}
 
   ngOnInit(): void {
     if(localStorage.getItem('isLoggedIn') === 'true') {
@@ -52,15 +65,15 @@ export class HeaderComponent implements OnInit {
     }
   }
 
-  handleLogoClick(): void {
+  handleLogo(): void {
     this.router.navigate(['/']);
   }
 
-  handleLoginClick(): void {
+  handleLogin(): void {
     this.auth.loginWithRedirect();
   }
 
-  handleLogoutClick(): void {
+  handleLogout(): void {
     const cartId = localStorage.getItem('cartId')?.toString();
     this.cartService.apiV1CartsCartIdDelete(cartId!).subscribe(
       (data: CartResponseDTO) => {
@@ -78,11 +91,34 @@ export class HeaderComponent implements OnInit {
     this.router.navigate(['/']);
   }
 
-  handleCartClick(): void {
+  handleCart(): void {
     this.router.navigate(['/cart']);
   }
 
-  handleNavigateToOrdersClick(): void {
+  handleNavigateToOrders(): void {
     this.router.navigate(['/orders']);
+  }
+
+  handleSearch(): void {
+    this.searchChanged.emit({
+      searchText: this.productListFilter.searchText,
+      category: this.productListFilter.category,
+      minPrice: this.productListFilter.minPrice,
+      maxPrice: this.productListFilter.maxPrice
+    });
+    this.productListFilter = {
+      category: this.productListFilter.category,
+      minPrice: this.productListFilter.minPrice,
+      maxPrice: this.productListFilter.maxPrice,
+      searchText: null
+    };
+  }
+
+  toggleCollapse(): void {
+    if (this.collapseElement.nativeElement.classList.contains('show')) {
+      this.renderer.removeClass(this.collapseElement.nativeElement, 'show');
+    } else {
+      this.renderer.addClass(this.collapseElement.nativeElement, 'show');
+    }
   }
 }