+++ frontend/e-commerce-service/src/app/components/product-management/product-list/product-list.component.ts
@@ -7,18 +7,28 @@ import { HeaderComponent } from '../../header/header.component';
 import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import { faPlus, faStar } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { ProductListFilter } from '../../../models/product-list-filter';
+import { FooterComponent } from '../../footer/footer.component';
 
 @Component({
   selector: 'app-product-list',
   standalone: true,
-  imports: [ CommonModule, FormsModule, FontAwesomeModule, HeaderComponent ],
+  imports: [ CommonModule, FormsModule, FontAwesomeModule, HeaderComponent, FooterComponent ],
   templateUrl: './product-list.component.html',
   styleUrls: ['./product-list.component.css']
 })
 export class ProductListComponent implements OnInit {
   @Input() products: ProductResponseDTO[] = [];
 
-  availableCategories: string[] = [
+  public productListFilter: ProductListFilter = {
+    category: null,
+    minPrice: null,
+    maxPrice: null,
+    searchText: null
+  };
+  public pageNumber: number = 0;
+
+  public availableCategories: string[] = [
     'Acoustic Guitar',
     'Electric Guitar',
     'Bass Guitar',
@@ -35,6 +45,8 @@ export class ProductListComponent implements OnInit {
   public faStar: IconDefinition = faStar;
   public isLoggedIn: boolean = false;
   public accessToken: string | null = ''; 
+
+  public enableSearch: boolean = true;
 
   constructor(private router: Router, private productService: ProductService, private cartService: CartService) {}
 
@@ -69,25 +81,52 @@ export class ProductListComponent implements OnInit {
         );
       }
     } 
-    this.productService.apiV1ProductsGet().subscribe(
+
+    const appendNewProducts: boolean = false;
+    this.getProducts(appendNewProducts);
+  }
+
+  handleCreateProduct(): void {
+    this.router.navigate(['/products/creation']);
+  }
+
+  handleViewProduct(product: ProductResponseDTO): void {
+    this.router.navigate(['/products', product.productID]);
+  }
+
+  handleViewReviews(product: ProductResponseDTO): void {
+    this.router.navigate(['/products', product.productID, 'reviews']);
+  }
+
+  handleSearchChanged(productListFilter: ProductListFilter): void {
+    this.productListFilter = productListFilter;
+    // console.log(this.productListFilter);
+    const appendNewProducts: boolean = false;
+    this.getProducts(appendNewProducts);
+  }
+
+  handleLoadMore(pageNumber: number): void {
+    this.pageNumber = pageNumber;
+    // console.log(this.pageNumber);
+    const appendNewProducts: boolean = true;
+    this.getProducts(appendNewProducts);
+  }
+
+  getProducts(append: boolean): void {
+    const pageSize = 3;
+    this.productService.apiV1ProductsGet(this.pageNumber, pageSize, this.productListFilter.category!, this.productListFilter.searchText!, this.productListFilter.minPrice!, this.productListFilter.maxPrice!).subscribe(
       (data: ProductResponseDTO[]) => {
-        this.products = data;
+        if(append) {
+          for(const d of data) {
+            this.products.push(d);
+          }
+        }  else {
+          this.products = data;
+        }
       },
       error => {
         console.error('Error fetching products', error);
       }
     );
-  }
-
-  handleCreateProductClick(): void {
-    this.router.navigate(['/products/creation']);
-  }
-
-  handleViewProductClick(product: ProductResponseDTO): void {
-    this.router.navigate(['/products', product.productID]);
-  }
-
-  handleViewReviewsClick(product: ProductResponseDTO): void {
-    this.router.navigate(['/products', product.productID, 'reviews']);
   }
 }