+++ frontend/e-commerce-service/src/app/components/review-management/review-creation/review-creation.component.ts
@@ -52,11 +52,11 @@ export class ReviewCreationComponent implements OnInit {
     } 
   }
 
-  handleNavigateBackClick(): void {
+  handleNavigateBack(): void {
     this.router.navigate(['/']);
   }
 
-  handleCreateReviewClick(): void {
+  handleCreateReview(): void {
     // create review
     const reviewRequestDto: ReviewRequestDTO = {
       productID: this.product.productID!,