+++ frontend/e-commerce-service/src/app/models/product-list-filter.ts
@@ -1,0 +1,7 @@
+import { BaseFilter } from "./base-filter";
+
+export interface ProductListFilter extends BaseFilter {
+    category: string | null;
+    minPrice: number | null;
+    maxPrice: number | null;
+}